quest app implementation version 05: auto formatting applied added readme with instruction how to deploy  war file on Tomcat Docker Container
