+++ Dockerfile
@@ -1,3 +1,4 @@
-FROM tomcat:9.0
+FROM tomcat:9-jdk17-corretto
+
 
 COPY ./target/javarush-quest.war /usr/local/tomcat/webapps/
