+++ src/main/webapp/WEB-INF/web.xml
@@ -1,13 +1,13 @@
 <!DOCTYPE web-app PUBLIC
- "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
- "http://java.sun.com/dtd/web-app_2_3.dtd" >
+        "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
+        "http://java.sun.com/dtd/web-app_2_3.dtd" >
 
 <web-app>
-  <display-name>Archetype Created Web Application</display-name>
-  <context-param>
-    <param-name>log4jConfiguration</param-name>
-    <param-value>/Users/alinakursekova/IdeaProjects/javarush-quest/src/main/resources/log4j2.xml</param-value>
-  </context-param>
+    <display-name>Archetype Created Web Application</display-name>
+    <context-param>
+        <param-name>log4jConfiguration</param-name>
+        <param-value>/Users/alinakursekova/IdeaProjects/javarush-quest/src/main/resources/log4j2.xml</param-value>
+    </context-param>
 
 </web-app>
